Refactoring updater.js
